Modify StopsController to use dependency injection instead of default string (#8)
* Add stations model class.

This will be used to represent a station and contains information that can be used for further lookups such as naptan ID.

* Add file that converts station names to TLAREFs.

This can be used to convert the station name to its 3 char code.

* Add file that converts TLAREFs to IDs.

These IDs can then be used to look up the next services at the ID. There are commonly multiple IDs at the same location so this needs to be taken into consideration.

* Add file that maps the TLAREF to the NaPTAN for the station.

* Rename TLAREFs_to_NaPTAN to be ATCOCode instead.

It is the ATCOCode that is used here instead of the NaPTAN value.

* Modify station to have 'AtcoCode' instead of NaptanId.

This is the correct terminology for this field. It is not a NaPTAN.

* Add Stops json file.

This contains information about all tram stops, such as Tlaref, IDs, location and nearest road crossings.

* Add line to stops data.

Each stop can only be on a single line. The Station model has been updated to show this aswell as matching the naming conventions.

* Add Open government license declaration

Add declaration of license for TfGM data.

* Rename Station to Stop.

This follows the new naming convention where Stop is used instead of Station.

* Remove old Station class.

* Add controller for '/api/stops'

This returns all stops as a JSON list. If the stops list is null, it will be imported from the stops file in resoruces.

* Add Newtonsof JSON NuGet package.

This is used for JSON Deserialisation when importing the stops.

* Add basic unit test project and update solution to include unit testing project.

* Reformat stops json to improve readability.

This splits the single line file into multiple lines.

* Add stop zone to stop data.

* Add StopZone to Stop model.

This represents the ticketing zone the stop falls into. If a stop is in multiple zones, it is represnted as a/b, where a is the lower of the two zones.

* Add reference to base tfgm api wrapper project.

This allows sections to be used in the test project.

* Create a new StopLoader class for loading in Stop related info.

This removes the method from the StopsController class.

* Move UnitTest1 to TestModels/TestStopsLoader.

This will be used to test the StopsLoader class.

* Add TestControllers folder.

This will be used for Controller testing files.

* Add test for creating a valid stop loader and get stop loader to use constructor path.

This should return a list that contains a single stop.

* Add test for creating StopLoader with null file path.

This throws an argument null exception.

* Modify StopLoader to check if the file given as a parmeter exists.

If not, a FileNotFoundException is thrown.

* Remove the test controllers directory from csproj.

This is no longer required in the csproj file.

* Add test for getting the correct status code from GetAllStops in StopsController.

This also adds a constructor for creating a custom controller with a custom file path.

* Modify StopsController to use default string instead of two constructors.

The use of multiple constructors is not supported.

* Add test to get default StopsController return code.

This should be a 200.

* Add test to check returned list length from GetAllStops.

This should be 1 as this is the number of stops in the file given to the StopLoader.

* Add test to test the default constructor value.

This should return 99 stops.

* Fix null warnings in TestStopsController.

* Move Tests project to dotnet 5.0 to match implementation project.

* Add dotnet test github action.

This will be used for running tests on pull requrests.

* Add StopsPath log to view the current working folder.

This is required to check why GitHub actions is failing.

* Remove logging from StopLoader and remove default controller tests.

The controller tests have been moved due to differences as to how dotnet is running on github actions.

* Add default route path to appSettings and import into configuration.

* Create a new Resources config POCO

This will be used for dependency injection for resources.

* Add ResourceConfig to StopsController and StopsController tests.

This is injected in when services are configured and by the tests.